requirements.txt
streamlit==1.30.0
praw==7.7.1
tweepy==4.14.0
requests==2.31.0
python-dotenv==1.0.0
emoji==2.10.0
